Style Animation updated to use the ReactSciChart
